Rename `CardPosition` to `CardLoc` and `position` to `loc`
`position.rotation` feels weird, and also I wanted it shorter.
Rotation feels better as part of a "location", and
"location" can be shortened to "loc".

I also considered "Box" or "OB" (for Oriented Box), but it's a bit
more bespoke than that implies since it's assumed to be a fixed size,
namely the size of the cards.